separate constants
