+++ .circleci/config.yml
@@ -49,10 +49,10 @@ jobs:
       - store_test_results:
           path: unit-tests-results.xml
 
-  e2e:
+  e2e_site:
     docker:
       - image: mcr.microsoft.com/playwright:v1.28.1-focal
-    parallelism: 8
+    parallelism: 4
     steps:
       - checkout
       - restore_cache:
@@ -71,7 +71,35 @@ jobs:
             - ~/.cache/yarn
       - run:
           name: E2E Test
-          command: SHARD="$((${CIRCLE_NODE_INDEX}+1))"; yarn test:e2e --shard=${SHARD}/${CIRCLE_NODE_TOTAL}
+          command: SHARD="$((${CIRCLE_NODE_INDEX}+1))"; yarn test:e2e --project="Site" --shard=${SHARD}/${CIRCLE_NODE_TOTAL}
+      - store_test_results:
+          path: results.xml
+      - store_artifacts:
+          path: test-results
+
+  e2e_glossary:
+    docker:
+      - image: mcr.microsoft.com/playwright:v1.28.1-focal
+    parallelism: 4
+    steps:
+      - checkout
+      - restore_cache:
+          name: Restore Yarn Package Cache
+          keys:
+            - yarn-packages-e2e-{{ .Branch }}-{{ checksum "yarn.lock" }}
+          paths:
+            - ~/.yarn/cache
+      - run:
+          name: Install Dependencies
+          command: yarn install --frozen-lockfile --cache-folder ~/.cache/yarn
+      - save_cache:
+          name: Save Yarn Package Cache
+          key: yarn-packages-e2e-{{ .Branch }}-{{ checksum "yarn.lock" }}
+          paths:
+            - ~/.cache/yarn
+      - run:
+          name: E2E Test
+          command: SHARD="$((${CIRCLE_NODE_INDEX}+1))"; yarn test:e2e --project="Glossary" --shard=${SHARD}/${CIRCLE_NODE_TOTAL}
       - store_test_results:
           path: results.xml
       - store_artifacts:
@@ -80,6 +108,7 @@ jobs:
 workflows:
   CI:
     jobs:
-      - unit_test
-      - e2e
-      - eslint
+      # - unit_test
+      - e2e_site
+      - e2e_glossary
+      # - eslint
