+++ e2e/site/navigation.spec.ts
@@ -1,13 +1,13 @@
 import { test, expect } from "@playwright/test";
 import { StylableDOMUtil } from "@stylable/dom-test-kit";
-import * as stylesheet from "../src/components/top-bar/top-bar.st.css";
-import * as headerStylesheet from "../src/components/header/header.st.css";
+import * as stylesheet from "../../src/components/top-bar/top-bar.st.css";
+import * as headerStylesheet from "../../src/components/header/header.st.css";
 import {
 	getLocalePath,
 	locales,
 	getFrontMatter,
 	translate,
-} from "./utils/test-utils";
+} from "../utils/test-utils";
 
 const domUtil = new StylableDOMUtil(stylesheet);
 const headerDomUtil = new StylableDOMUtil(headerStylesheet);
@@ -15,7 +15,7 @@ const headerDomUtil = new StylableDOMUtil(headerStylesheet);
 const topBar = domUtil.scopeSelector(".root");
 const siteTitle = headerDomUtil.scopeSelector(".siteTitle");
 
-const docId = "docs/the-story-of-mel";
+// const docId = "docs/the-story-of-mel";
 
 test.describe("TopBar Navigation", () => {
 	locales.map((locale) => {
@@ -57,26 +57,6 @@ test.describe("Dynamic Pages", () => {
 			await page.hover(`text=${translate(locale, "MENU_ITEM_LABEL_ID_ABOUT")}`);
 			await page.click(
 				`text=${translate(locale, "MENU_ITEM_LABEL_EXCERPT_SHOW_MORE")}`
-			);
-
-			await expect(page).toHaveURL(localePath);
-			await expect(page.locator("h1")).toHaveText(data.title as string);
-		});
-
-		test(`${locale} > should navigate to the Resources page`, async ({
-			page,
-		}) => {
-			const path = "pages/resources";
-			const filename = "index";
-			const { data } = getFrontMatter(docId, `${path}/${filename}`, locale);
-			const localePath = getLocalePath(locale, docId, path);
-
-			await page.goto(localePath);
-			await page.hover(
-				`text=${translate(locale, "MENU_SECTION_LABEL_ARTICLES")}`
-			);
-			await page.click(
-				`text=${translate(locale, "MENU_ITEM_LABEL_ID_RESOURCES")}`
 			);
 
 			await expect(page).toHaveURL(localePath);
